ARM: dts: 14001: Fix overexposure when capturing photos with flash
When photos are captured with flash, there are 2 phases of the snapshot:
pre-flash (low power) and the actual flash (high power). During the low-power
flash phase, 3A acquires data on the lighting and uses it to configure the
actual snapshot with high-power flash. Since the gap between low-power and
high-power flash is so great (53 mA versus 560 mA, respectively), 3A
overcompensates exposure by assuming the scene with high-power flash is much
darker than it actually is. This results in the final photo with high-power
flash being severely overexposed and susceptible to motion blur due to the use
of a longer-than-necessary exposure time.

By increasing the current of the LEDs for torch-mode from 53 mA to 130 mA,
the pre-flash phase when capturing a photo appears brighter to 3A, so 3A
won't overestimate the exposure. Photos with flash are now properly exposed.

130 mA was derived empirically using a pitch-black room as a the control
variable.

Additionally, the torch is now ~2.5x brighter as a result.

Change-Id: I9a67b111b04f9ce97ab05fae50e5bd377758eb10
Signed-off-by: Sultanxda <sultanxda@gmail.com>